refactor: App moves to components
